fix(dashboard): change production_total_month schema type to float and handle None values in goal overall achievement percent query

+++ backend/app/modules/dashboard/schemas.py
@@ -25,5 +25,5 @@ class DashboardMetrics(BaseModel):
     efficiency_label: str = "Operacional"  # "Operacional" | "Crítico" | "Excelente"
 
     # Dados adicionais para gráficos
-    production_total_month: int = 0
+    production_total_month: float = 0.0
     production_growth_pct: float | None = None
